refactor(#98): step 별로 컴포넌트 나눔
