+++ apps/web/src/components/signup/SignUpContent/steps/UsernameStep.tsx
@@ -1,0 +1,60 @@
+import styled from '@emotion/styled';
+import { tokens } from '@horizon/tokens';
+import { Text, TextField } from '@horizon/ui';
+import type { UsernameStepProps } from '../types';
+
+const UsernameStep = ({
+  username,
+  setUsername,
+  onSubmit,
+  isLoading,
+  errors,
+}: UsernameStepProps) => {
+  return (
+    <>
+      <TextField
+        label='이름'
+        placeholder='이름 입력'
+        type='text'
+        width='100%'
+        value={username}
+        onChange={(e) => setUsername(e.target.value)}
+        error={!!errors.username}
+        helperText={errors.username}
+      />
+
+      <StyledSubmitButton
+        onClick={onSubmit}
+        disabled={!username.trim() || !!errors.username || isLoading}
+        style={{ marginTop: '60px' }}
+      >
+        <Text variant='ST' color={tokens.colors.white}>
+          {isLoading ? '완료 중...' : '완료'}
+        </Text>
+      </StyledSubmitButton>
+    </>
+  );
+};
+
+const StyledSubmitButton = styled.button<{ disabled: boolean }>`
+  width: 100%;
+  height: 48px;
+  background-color: ${({ disabled }) => (disabled ? tokens.colors.neutral[300] : tokens.colors.primary[500])};
+  border: none;
+  border-radius: ${tokens.rounding.object};
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: ${({ disabled }) => (disabled ? 'not-allowed' : 'pointer')};
+  transition: background-color 0.2s ease;
+  
+  &:hover {
+    background-color: ${({ disabled }) => (disabled ? tokens.colors.neutral[300] : tokens.colors.primary[600])};
+  }
+  
+  &:disabled {
+    opacity: 1;
+  }
+`;
+
+export default UsernameStep;
